Fix #33 - reduce command failure when start not zero
The commands used for select and reduce are provided with both animation data and a list of keyframes. Notably, the animation data is collected by the Python script running the interface. To reduce memory and any operations for trimming, I pass only the relevant animation data to these commands. While this the benefits of no trimming needed, it does have the one consequence that the list of keyframes provided to those commands must start from zero (since the animation data only includes what is relevant).

I had accounted for the above in the previous GUI code, but forgot that I needed to account for this starting-from-zero when actually processing the reduction result (deleting old keyframes and keying in new keyframes). Here I have added the lines to properly handle this problem here.