Add check to report when fixed keyframes are out of bounds
This new code reports error whenever the value of a fixed keyframes is greater or less than the start/end frame for selection. While I could just ignore the out-of-bounds frames, I expect its better to notify artists of this in case they don't realise that the fixed keyframe will be excluded.