Interface redesign - linear layout and better visuals
This is a somewhat turblent change where a number of things came together at once. First I added a divider, that just separated the interface into different areas. I went back and forth with a number of options, but in the end I grouped together a "fixed frames" area at the top, and then a "selection" area and a "reduce area" lower down. These areas represent the linear workflow of using the tool: (1) define where to choose and say what is fixed, (2) select extremes and breakdowns, and (3) apply the reduction.

Some key things:

- I no longer use a start/end frame explicitly. Instead I just have the start frame being the first fixed frame, and the last fixed being the end.

- Fixed keyframes are presented with a table interface. This is more expressive in that I can support adding names and notes for keyframes, and I believe that the table version will help to reduce confusion about setting fixed keyframes.

- I've added locks to disable the keyframe sliders until the select buttons are pressed (these get locked again if the fixed keyframes change).

- I've split up the attributes used for extremes and breakdowns. Keeping two channels will reduce rework if an artist is going back and forth with the tool.

- I added a new button to "lock in" the extremes. This is cosemetic only, but is helpful in that the artist can choose to duplicate an object at all extreme keyframes. This gives a visual point of difference when viewing potential keyframes.

Okay, I think that's about it. Somewhat of a mess :)